+++ lib/EnumTrait.php
@@ -1,10 +1,10 @@
 <?php
-namespace cedx;
+namespace enum;
 
 /**
  * Provides static methods for enumerations.
  */
-trait Enum {
+trait EnumTrait {
 
   /**
    * Private constructor: prohibit the class instantiation.